requirements.txt
PyQt6==6.4.0
numpy==1.24.1
pytest-qt==4.2.0
